// client/src/components/review/IReviewComponentState.ts
export interface IReviewComponentState {
  /**
   * subtotal of the cart products
   * @type {number}
   * @memberof IReviewComponentState
   */
  subtotalofcart?: number

  /**
   * Grandtotal of the cart products
   * @type {number}
   * @memberof IReviewComponentState
   */
  grandtotalofcart?: number
}